Moving _unused macro to _unused_attr
Currently xbps package fails to build using musl libc because the signal.h
header includes a struct field called _unused, that becomes defined as an
function attribute due to the a CFLAG -D macro definition. This cause the
build to fail.

This patch just rename _unused macro to _unused_attr in order to avoid the
names being clashed.